+++ app/src/main/java/de/igloffstein/maik/arevelation/adapters/RevelationStructureBrowserAdapter.java
@@ -2,6 +2,7 @@ package de.igloffstein.maik.arevelation.adapters;
 
 import android.app.Activity;
 import android.support.annotation.NonNull;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -9,6 +10,7 @@ import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.github.marmaladesky.ARevelation;
 import com.github.marmaladesky.R;
 import com.github.marmaladesky.data.Entry;
 
@@ -24,14 +26,15 @@ import de.igloffstein.maik.arevelation.enums.EntryType;
 
 public class RevelationStructureBrowserAdapter extends ArrayAdapter<String> {
 
-    private final Activity context;
+    private static final String LOG_TAG = RevelationStructureBrowserAdapter.class.getSimpleName();
+    private final ARevelation aRevelation;
     private final List<Entry> data;
 
-    public RevelationStructureBrowserAdapter(Activity context, List<Entry> data) {
+    public RevelationStructureBrowserAdapter(Activity activity, List<Entry> data) {
         //noinspection unchecked
-        super(context, R.layout.revelation_structure_entry, (List) data);
+        super(activity, R.layout.revelation_structure_entry, (List) data);
 
-        this.context = context;
+        this.aRevelation = (ARevelation) activity;
         this.data = data;
     }
 
@@ -44,58 +47,82 @@ public class RevelationStructureBrowserAdapter extends ArrayAdapter<String> {
      */
     @Override
     @NonNull
-    public View getView(int position, View view, @NonNull ViewGroup parent) {
-        LayoutInflater inflater = context.getLayoutInflater();
-        View rowView = inflater.inflate(R.layout.revelation_structure_entry, parent, false);
+    public View getView(final int position, View view, @NonNull ViewGroup parent) {
+        LayoutInflater inflater = aRevelation.getLayoutInflater();
+        final View rowView = inflater.inflate(R.layout.revelation_structure_entry, parent, false);
 
-        ImageView imageView = rowView.findViewById(R.id.imageIdentifier);
-        TextView txtTitle = rowView.findViewById(R.id.headerIdentifier);
+        ImageView entryTypeIcon = rowView.findViewById(R.id.imageIdentifier);
+        TextView entryName = rowView.findViewById(R.id.headerIdentifier);
+        ImageView entryDeleteIcon = rowView.findViewById(R.id.deleteImageView);
 
-        if (data != null && data.get(position) != null) {
-            txtTitle.setText(data.get(position).getName());
-            switch (EntryType.valueOf(data.get(position).type.toUpperCase())) {
-                case CREDITCARD:
-                    imageView.setImageResource(R.drawable.ic_cc_938);
-                    break;
-                case CRYPTOKEY:
-                    imageView.setImageResource(R.drawable.ic_key_678);
-                    break;
-                case DATABASE:
-                    imageView.setImageResource(R.drawable.ic_database_system_1800);
-                    break;
-                case DOOR:
-                    imageView.setImageResource(R.drawable.ic_door_44);
-                    break;
-                case EMAIL:
-                    imageView.setImageResource(R.drawable.ic_email_1573);
-                    break;
-                case FTP:
-                    imageView.setImageResource(R.drawable.ic_object_connection_round_1096);
-                    break;
-                case GENERIC:
-                    imageView.setImageResource(R.drawable.ic_attachment_1574);
-                    break;
-                case PHONE:
-                    imageView.setImageResource(R.drawable.ic_phone_225);
-                    break;
-                case SHELL:
-                    imageView.setImageResource(R.drawable.ic_window_equal_1465);
-                    break;
-                case REMOTEDESKTOP:
-                    imageView.setImageResource(R.drawable.ic_screen_1111);
-                    break;
-                case VNC:
-                    imageView.setImageResource(R.drawable.ic_showcase_round_726);
-                    break;
-                case WEBSITE:
-                    imageView.setImageResource(R.drawable.ic_file_url_1759);
-                    break;
-                case FOLDER:
-                    imageView.setImageResource(R.drawable.ic_folder_1791);
-                    break;
-                default:
+        if (data != null) {
+            final Entry entry = data.get(position);
+            if (entry != null) {
+
+                /**
+                 * [TypeIcon] [EntryName]       [EntryDeleteIcon]
+                 */
+                
+                // set the icon to display
+                entryTypeIcon.setImageResource(getIconForView(EntryType.valueOf(entry.type.toUpperCase())));
+               
+                // read name of item and display it (text)
+                entryName.setText(entry.getName());
+
+                // set the delete function
+                entryDeleteIcon.setOnClickListener(new View.OnClickListener() {
+
+                    @Override
+                    public void onClick(View v) {
+                        Log.d(LOG_TAG, "Deleting: " + entry.getUuid());
+                        // remove from the main data storage
+                        ((ARevelation) aRevelation).getRvlData().removeEntryById(entry.getUuid());
+                        // remove from the adapter view
+                        data.remove(entry);
+                        // notify data changes
+                        notifyDataSetChanged();
+                        // display save button
+                        aRevelation.getSaveButton().setVisibility(View.VISIBLE);
+                    }
+                });
+
+                
             }
         }
         return rowView;
+    }
+    
+    private Integer getIconForView(EntryType entryType){
+        // witch icon shall we display
+        switch (entryType) {
+            case CREDITCARD:
+                return R.drawable.ic_cc_938;
+            case CRYPTOKEY:
+                return R.drawable.ic_key_678;
+            case DATABASE:
+                return R.drawable.ic_database_system_1800;
+            case DOOR:
+                return R.drawable.ic_door_44;
+            case EMAIL:
+                return R.drawable.ic_email_1573;
+            case FTP:
+                return R.drawable.ic_object_connection_round_1096;
+            case GENERIC:
+                return R.drawable.ic_attachment_1574;
+            case PHONE:
+                return R.drawable.ic_phone_225;
+            case SHELL:
+                return R.drawable.ic_window_equal_1465;
+            case REMOTEDESKTOP:
+                return R.drawable.ic_screen_1111;
+            case VNC:
+                return R.drawable.ic_showcase_round_726;
+            case WEBSITE:
+                return R.drawable.ic_file_url_1759;
+            case FOLDER:
+                return R.drawable.ic_folder_1791;
+            default:
+                return null;
+        }
     }
 }